Change IV to account for discreate process
Exponential distribution is defined as `dN/dt = -λN` but that formula is correct for continuos process.
In the case of a discrete process (our case), we have to work from the solution of the differential equation:
```
N(t) = N₀*e^(-λ*t)
IV = N(0) - N(1)
IV = N₀ - N₀*e^(-λ)
IV = N₀ * (1-e^(-λ))
(1-e^(-λ)) = IV/N₀  (eq. 1)
```
Then the reward is:
```
R = C(0) - C(1)
R = C₀ - C₀*e^(-λ)
R = C₀ * (1-e^(-λ)) // substitute eq. 1
R = (C₀ * IV) / N₀
```
Where `C` is NetworkActorBalance.

Above means that the `Reward` equation is the same, but the IV equation is different.
The previous IV was causing the reward schedule to be 26FIL out at the six years mark (halving point).
With the new IV, the schedule is within 3.94311e-12 at the same point in time (as implemented in a Filecoin client).
The source of an observed error is limited precision of FIL unit (attoFIL) but is small enough that it should not matter.